+++ src/Models/Installation.php
@@ -2,6 +2,7 @@
 
 namespace Agnes\Models;
 
+use Agnes\Models\Installation\OnlinePeriod;
 use DateTime;
 
 class Installation
@@ -17,9 +18,9 @@ class Installation
     private $number;
 
     /**
-     * @var Setup
+     * @var string
      */
-    private $setup;
+    private $releaseOrCommitish;
 
     /**
      * @var OnlinePeriod[]
@@ -29,11 +30,11 @@ class Installation
     /**
      * Installation constructor.
      */
-    public function __construct(string $folder, int $number, Setup $setup, array $onlinePeriods = [])
+    public function __construct(string $folder, int $number, string $releaseOrCommitish, array $onlinePeriods = [])
     {
         $this->folder = $folder;
         $this->number = $number;
-        $this->setup = $setup;
+        $this->releaseOrCommitish = $releaseOrCommitish;
         $this->onlinePeriods = $onlinePeriods;
     }
 
@@ -47,9 +48,9 @@ class Installation
         return $this->number;
     }
 
-    public function getSetup(): Setup
+    public function getReleaseOrCommitish(): string
     {
-        return $this->setup;
+        return $this->releaseOrCommitish;
     }
 
     /**
@@ -84,7 +85,7 @@ class Installation
 
     public function toArray(): array
     {
-        $array = ['number' => $this->number, 'setup' => $this->setup->toArray(), 'online_periods' => []];
+        $array = ['number' => $this->number, 'release_or_commitish' => $this->releaseOrCommitish, 'online_periods' => []];
 
         foreach ($this->onlinePeriods as $onlinePeriod) {
             $array['online_periods'][] = $onlinePeriod->toArray();
@@ -95,12 +96,11 @@ class Installation
 
     public static function fromArray(string $folder, array $array): Installation
     {
-        $setup = Setup::fromArray($array['setup']);
         $onlinePeriods = [];
         foreach ($array['online_periods'] as $onlinePeriod) {
             $onlinePeriods[] = OnlinePeriod::fromArray($onlinePeriod);
         }
 
-        return new Installation($folder, $array['number'], $setup, $onlinePeriods);
+        return new Installation($folder, $array['number'], $array['release_or_commitish'], $onlinePeriods);
     }
 }
