+++ src/Services/Policy/DeployPolicyVisitor.php
@@ -61,7 +61,7 @@ class DeployPolicyVisitor extends PolicyVisitor
         // check if the release was published there at any given time
         foreach ($instances as $instance) {
             foreach ($instance->getInstallations() as $installation) {
-                if ($installation->getSetup()->getIdentification() === $this->deployment->getSetup()->getIdentification()) {
+                if ($installation->getReleaseOrCommitish() === $this->deployment->getSetup()->getIdentification()) {
                     return true;
                 }
             }
