+++ src/Services/Task/Payload/Release.php
@@ -3,20 +3,13 @@
 namespace Agnes\Actions;
 
 use Agnes\Actions\Visitors\AbstractActionVisitor;
-use Agnes\Services\PolicyService;
-use Exception;
 
 class Release extends AbstractPayload
 {
     /**
-     * @var string|null
-     */
-    private $name;
-
-    /**
      * @var string
      */
-    private $commitish;
+    private $name;
 
     /**
      * Release constructor.
@@ -25,7 +18,6 @@ class Release extends AbstractPayload
      */
     public function __construct(string $commitish, string $name = null)
     {
-        $this->commitish = $commitish;
         $this->name = $name;
     }
 
@@ -39,17 +31,9 @@ class Release extends AbstractPayload
         return null !== $this->name ? $this->name : $this->commitish;
     }
 
-    /**
-     * @throws Exception
-     */
-    public function canExecute(PolicyService $policyService): bool
-    {
-        return $policyService->canRelease($this);
-    }
-
     public function describe(): string
     {
-        return 'build '.$this->getCommitish();
+        return 'release '.$this->getName();
     }
 
     public function accept(AbstractActionVisitor $abstractActionVisitor): bool
